+++ Ristorante/src/main/java/model/Pagamento.java
@@ -1,15 +1,13 @@
 package model;
 
-import java.math.BigDecimal;
-
 public class Pagamento {
 	private int id_pagamento;
 	private int id_tavolo;
-	private BigDecimal costo_totale;
+	private float costo_totale;
 	private String stato;
 	private String data;
 
-	public Pagamento(int id,int id_tavolo, BigDecimal ct, String s, String d) {
+	public Pagamento(int id,int id_tavolo, float ct, String s, String d) {
 		this.id_pagamento=id;
 		this.id_tavolo = id_tavolo;
 		this.costo_totale = ct;
@@ -36,11 +34,11 @@ public class Pagamento {
 		this.id_tavolo = id_tavolo;
 	}
 
-	public BigDecimal getCosto_totale() {
+	public float getCosto_totale() {
 		return costo_totale;
 	}
 
-	public void setCosto_totale(BigDecimal costo_totale) {
+	public void setCosto_totale(float costo_totale) {
 		this.costo_totale = costo_totale;
 	}
